Changed return type of `mt19937_64_rand_real` to `double long`.
Clang pointed out that converting 0xFFFFFFFFFFFFFFFFU to `double` changes its value from 18446744073709551615 to 18446744073709551616. This does not happen if it is converted to `double long` when using GCC or Clang.

+++ include/mt19937.h
@@ -48,7 +48,7 @@ double mt19937_rand_real(struct mt19937_t *mt);
 #define mt19937_rand_real(...)  \
 mt19937_rand_real(GET_OR_NULL(__VA_ARGS__ __VA_OPT__(,) NULL))
 
-double mt19937_64_rand_real(struct mt19937_64_t *mt);
+double long mt19937_64_rand_real(struct mt19937_64_t *mt);
 #define mt19937_64_rand_real(...)  \
 mt19937_64_rand_real(GET_OR_NULL(__VA_ARGS__ __VA_OPT__(,) NULL))
